+++ src/components/App.jsx
@@ -3,7 +3,7 @@ import { Switch } from "react-router-dom";
 import ConditionRoute from "./general/conditionRoute";
 
 import Login from "./login/Login";
-import Message from "./message/Message";
+import Messenger from "./messenger/Messenger";
 
 import "../css/messageApp/app/App.css";
 
@@ -20,7 +20,7 @@ const App = () => (
         condition={{ loggedOut: true, route: "/login" }}
         exact
         path="/user"
-        component={Message}
+        component={Messenger}
       />
     </Switch>
   </div>
