+++ src/components/messenger/message/messageSend/MessageSend.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../../css/messageApp/message/messageSend/messageSend.css";
+import "../../../../css/messageApp/message/messageSend/messageSend.css";
 
 // Components
 import MessageSendMessage from "./messageType/MessageSendMessage";
