+++ src/components/messenger/message/messageSend/messageType/MessageSendEmoji.jsx
@@ -2,7 +2,6 @@ import React from "react";
 import ProptTypes from "prop-types";
 import { connect } from "react-redux";
 // Actions
-import { toggleEmoji } from "../../../../actions/Actions";
 
 const MessageSendSmile = props => (
   <div
@@ -18,7 +17,4 @@ MessageSendSmile.propTypes = {
   toggleEmoji: ProptTypes.func.isRequired
 };
 
-export default connect(
-  null,
-  { toggleEmoji }
-)(MessageSendSmile);
+export default MessageSendSmile;
