+++ src/components/messenger/sidebar/Sidebar.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import "../../../css/messageApp/sidebar/sidebar.css";
+
+const Sidebar = () => (
+  <div className="sidebar-wrap">
+    <div />
+  </div>
+);
+
+export default Sidebar;
