--- src/components/sidebarMenu/SidebarMessageItem.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-const SideBarMessageItem = () => <div className="sidebar_message">First</div>;
-
-export default SideBarMessageItem;
