+++ src/css/messageApp/messenger/messenger.css
@@ -1,0 +1,4 @@
+.messenger-wrap {
+  display: flex;
+  width: 100%;
+}
