+++ src/css/messageApp/sidebar/sidebar.css
@@ -1,67 +1,10 @@
-.sidebar {
-  height: 100vh;
+.sidebar-wrap {
+  height: 100%;
   width: 70px;
-  background: rgba(0, 91, 234, 0.85);
+  background: rgb(253, 253, 253);
+  border: 1px solid gray;
   flex-shrink: 0;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-}
-
-.sidebar_item {
-  margin: 10px 0;
-  height: 45px;
-  width: 45px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-}
-
-.sidebar_user_main {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-.sidebar_options {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.sidebar_logout_main {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-.sidebar_user_pic {
-  box-sizing: border-box;
-  height: 100%;
-  width: 100%;
-  border-radius: 50%;
-  background: url(https://picsum.photos/100);
-  background-size: contain;
-  background-position: center center;
-  border: 4px solid white;
-  margin-top: 20px;
-}
-
-/* Menus  */
-
-.sidebar_menu {
-  height: 100vh;
-  width: 300px;
-  background: rgb(250, 250, 250);
-  position: absolute;
-  z-index: 6000;
-  left: 70px;
-  top: 0;
 }
