--- src/css/messageApp/sidebar/sidebarSearch.css
@@ -1,3 +1,0 @@
-.sidebar_menu_search{
-  
-}
